Added el6 signer for a pre-baked AMI
This simplifies the script by removing the call out to erb
and not setting up the repos.

It also adds a neat bit of functionality which requires the
EC2 CLI utils (and java) to be installed: if you have tag
your instances with a tag named 'role' that describes what
they ought to do, we'll embed that tag's value into the cert,
under a new OID for easy classification.